chore: complete product-only rename — no repo agent harness leftovers
Rename source to host/runtime vocabulary:
- host-agents-md, host-claude-md (write AGENTS.md/CLAUDE.md into *user* projects)
- runtime-claude (Claude Code host adapter)
- agent-info types, agents-md-discovery
- templates/crew/roles/* (was crew/agents)
- scripts/install-standalone.sh (was install-via-claude)

Repo tree remains free of root AGENTS/CLAUDE/copilot/.claude/.mcp.
