+++ core/__tests__/services/host-agents-md.test.ts
@@ -12,7 +12,7 @@ import { afterEach, beforeEach, describe, expect, it } from 'bun:test'
 import fs from 'node:fs/promises'
 import os from 'node:os'
 import path from 'node:path'
-import { _routing, writeProjectAgentsMd } from '../../services/project-agents-md'
+import { _routing, writeProjectAgentsMd } from '../../services/host-agents-md'
 
 let dir: string
