+++ core/__tests__/services/host-claude-md.test.ts
@@ -15,7 +15,7 @@ import { afterEach, beforeEach, describe, expect, it } from 'bun:test'
 import fs from 'node:fs/promises'
 import os from 'node:os'
 import path from 'node:path'
-import { _routing, writeProjectClaudeMd } from '../../services/project-claude-md'
+import { _routing, writeProjectClaudeMd } from '../../services/host-claude-md'
 
 let dir: string
