+++ core/types/agent-info.ts
